+++ blockchain/chain/src/store/chain_store.rs
@@ -1,28 +1,24 @@
 // Copyright 2019-2023 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use std::{collections::VecDeque, num::NonZeroUsize, path::Path, sync::Arc, time::SystemTime};
+use std::{num::NonZeroUsize, path::Path, sync::Arc, time::SystemTime};
 
 use ahash::{HashMap, HashMapExt, HashSet};
 use anyhow::Result;
 use async_stream::stream;
 use bls_signatures::Serialize as SerializeBls;
-use cid::{
-    multihash::{Code, Code::Blake2b256},
-    Cid,
-};
+use cid::{multihash::Code::Blake2b256, Cid};
 use digest::Digest;
 use forest_beacon::{BeaconEntry, IGNORE_DRAND_VAR};
 use forest_blocks::{Block, BlockHeader, FullTipset, Tipset, TipsetKeys, TxMeta};
 use forest_interpreter::BlockMessages;
-use forest_ipld::{recurse_links_hash, CidHashSet};
+use forest_ipld::{should_save_block_to_snapshot, walk_snapshot};
 use forest_libp2p_bitswap::{BitswapStoreRead, BitswapStoreReadWrite};
 use forest_message::{ChainMessage, Message as MessageTrait, SignedMessage};
 use forest_metrics::metrics;
 use forest_networks::ChainConfig;
 use forest_shim::{
     address::Address,
-    clock::EPOCHS_IN_DAY,
     crypto::{Signature, SignatureType},
     econ::TokenAmount,
     executor::Receipt,
@@ -36,7 +32,6 @@ use forest_utils::{
     },
     io::Checksum,
 };
-use futures::Future;
 use fvm_ipld_amt::Amtv0 as Amt;
 use fvm_ipld_blockstore::Blockstore;
 use fvm_ipld_car::CarHeader;
@@ -562,7 +557,7 @@ where
 
         // Walks over tipset and historical data, sending all blocks visited into the
         // car writer.
-        Self::walk_snapshot(tipset, recent_roots, |cid| {
+        walk_snapshot(tipset, recent_roots, |cid| {
             let tx_clone = tx.clone();
             async move {
                 let block = self
@@ -570,15 +565,10 @@ where
                     .get(&cid)?
                     .ok_or_else(|| Error::Other(format!("Cid {cid} not found in blockstore")))?;
 
-                // Don't include identity CIDs.
-                // We only include raw and dagcbor, for now.
-                // Raw for "code" CIDs.
-                if u64::from(Code::Identity) != cid.hash().code()
-                    && (cid.codec() == fvm_shared::IPLD_RAW
-                        || cid.codec() == fvm_ipld_encoding::DAG_CBOR)
-                {
+                if should_save_block_to_snapshot(&cid) {
                     tx_clone.send_async((cid, block.clone())).await?;
                 }
+
                 Ok(block)
             }
         })
@@ -603,60 +593,6 @@ where
 
         let digest = writer.lock().await.get_mut().finalize();
         Ok(digest)
-    }
-
-    /// Walks over tipset and state data and loads all blocks not yet seen.
-    /// This is tracked based on the callback function loading blocks.
-    pub async fn walk_snapshot<F, T>(
-        tipset: &Tipset,
-        recent_roots: ChainEpoch,
-        mut load_block: F,
-    ) -> Result<(), Error>
-    where
-        F: FnMut(Cid) -> T + Send,
-        T: Future<Output = Result<Vec<u8>, anyhow::Error>> + Send,
-    {
-        let mut seen = CidHashSet::default();
-        let mut blocks_to_walk: VecDeque<Cid> = tipset.cids().to_vec().into();
-        let mut current_min_height = tipset.epoch();
-        let incl_roots_epoch = tipset.epoch() - recent_roots;
-
-        while let Some(next) = blocks_to_walk.pop_front() {
-            if !seen.insert(&next) {
-                continue;
-            }
-
-            let data = load_block(next).await?;
-
-            let h = BlockHeader::unmarshal_cbor(&data)?;
-
-            if current_min_height > h.epoch() {
-                current_min_height = h.epoch();
-                if current_min_height % EPOCHS_IN_DAY == 0 {
-                    info!(target: "chain_api", "export at: {}", current_min_height);
-                }
-            }
-
-            if h.epoch() > incl_roots_epoch {
-                recurse_links_hash(&mut seen, *h.messages(), &mut load_block).await?;
-            }
-
-            if h.epoch() > 0 {
-                for p in h.parents().cids() {
-                    blocks_to_walk.push_back(*p);
-                }
-            } else {
-                for p in h.parents().cids() {
-                    load_block(*p).await?;
-                }
-            }
-
-            if h.epoch() == 0 || h.epoch() > incl_roots_epoch {
-                recurse_links_hash(&mut seen, *h.state_root(), &mut load_block).await?;
-            }
-        }
-
-        Ok(())
     }
 }