+++ CHANGELOG.md
@@ -21,7 +21,10 @@ Notable updates:
 - [forest daemon] Support for NV18.
   [#2558](https://github.com/ChainSafe/forest/pull/2558)
   [#2579](https://github.com/ChainSafe/forest/pull/2579)
+- [forest daemon] Automatic database garbage collection.
+  [#2638](https://github.com/ChainSafe/forest/pull/2638)
 - [forest daemon] Add `--track-peak-rss` to forest daemon
+  [#2696](https://github.com/ChainSafe/forest/pull/2696)
 
 ### Changed